hide some private structs
